Summary
This application was created for uni project. The concept was; to develop a software product that provide a NEW service that is not already on the market.  The product needs to help people or resolves problem relating to EDUCATION.

This was done as a group assignment for Agile and lean Software Certification.